% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{biodbLipidmaps-package}
\alias{biodbLipidmaps}
\alias{biodbLipidmaps-package}
\title{biodbLipidmaps: biodbLipidmaps, a library for connecting to the Lipidmaps Structure database}
\description{
The biodbLipidmaps library provides access to the Lipidmaps Structure Database, using biodb package framework. It allows to retrieve entries by their accession number, and run web the services lmsdSearch and lmsdRecord.
}
\details{
See vignette intro:
```
vignette('intro', package='biodbLipidmaps')
```
}
\seealso{
\link{LipidmapsStructureConn}.
}
\author{
\strong{Maintainer}: Pierrick Roger \email{pierrick.roger@cea.fr}

}
